Reorganized the components.
