+++ src/features/cats/presentation/pages/CatBreedDetails.tsx
@@ -1,8 +1,9 @@
-import React, {useContext} from 'react';
+import React, { useContext } from 'react';
 import { useHistory, useParams } from 'react-router-dom';
 import { Button, Card, Container } from "react-bootstrap";
-import Breed from "../../data/models/Breed";
-import { SelectedBreedContext } from "../../providers/BreedsProvider";
+
+import Breed from '../../data/models/Breed';
+import { SelectedBreedContext } from '../../providers/BreedsProvider';
 
 interface Params {
   id: string
